Type fix in spirodraw
The "I'm feeling lucky" button hits this in dartium.

Review URL: http://codereview.chromium.org//8872005

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@2281 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/client/samples/spirodraw/ColorPicker.dart
@@ -31,7 +31,7 @@ class ColorPicker {
   String get selectedColor() => _selectedColor;
   
   void set selectedColor(num i) {
-    _selectedColor = getHexString(i.floor());
+    _selectedColor = getHexString(i.floor().toInt());
     showSelected();
     fireSelected();
   }
